Tidy lib: collapse utils/, move models to device/
Merge the lone src/lib/utils.ts into the utils/ folder (cn.ts + an index
barrel so @/lib/utils keeps resolving), and move the Portal model metadata
to src/lib/device/models.ts. No behavior change.
