+++ src/lib/device/models.ts
